+++ packages/medusa/src/services/__tests__/swap.js
@@ -742,6 +742,13 @@ describe("SwapService", () => {
         },
       }
 
+      const cartService = {
+        update: jest.fn(),
+        withTransaction: function() {
+          return this
+        },
+      }
+
       const swapRepo = MockRepository({
         findOneWithRelations: () => Promise.resolve(existing),
       })
@@ -753,6 +760,7 @@ describe("SwapService", () => {
         lineItemService,
         eventBusService,
         fulfillmentService,
+        cartService,
       })
 
       it("creates a shipment", async () => {
@@ -832,11 +840,23 @@ describe("SwapService", () => {
       },
     }
 
+    const cartService = {
+      update: () => {
+        return Promise.resolve()
+      },
+      withTransaction: function() {
+        return this
+      },
+    }
+
     const paymentProviderService = {
       getStatus: jest.fn(() => {
         return Promise.resolve("authorized")
       }),
       updatePayment: jest.fn(() => {
+        return Promise.resolve()
+      }),
+      cancelPayment: jest.fn(() => {
         return Promise.resolve()
       }),
       withTransaction: function() {
@@ -873,6 +893,7 @@ describe("SwapService", () => {
         eventBusService,
         swapRepository: swapRepo,
         totalsService,
+        cartService,
         paymentProviderService,
         eventBusService,
         shippingOptionService,
@@ -934,6 +955,7 @@ describe("SwapService", () => {
         eventBusService,
         swapRepository: swapRepo,
         totalsService,
+        cartService,
         paymentProviderService,
         eventBusService,
         shippingOptionService,